+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [1.0.2] - 2022-03-... (unreleased)
+### Changed
+- changes `OnnxInferenceModel` in `mltu.torch.inferenceModels` to load custom metadata from saved ONNX model
+- improved `mltu.dataProvider` to remove bad samples from dataset on epoch end
+
+### Added:
+- added `mltu.torch.losses`, used to create PyTorch losses, that may be used in training and validation
+- added CTC loss to `mltu.torch.losses` that can be used for training CTC based models
+- added `Model2onnx` and `Tensorboard` callbacks to `mltu.torch.callbacks`, used to create PyTorch callbacks, that may be used in training and validation
+- added `CERMetric` and `WERMetric` to `mltu.torch.metrics`, used to create PyTorch metrics, that may be used in training and validation
+- created 08 pytorch tutorial, that shows how to use `mltu.torch` to train CTC based models
+
+
 ## [1.0.1] - 2022-03-06
 ### Changed
 - In all tutorials removed stow dependency and replaced with os package, to make it easier to use on Windows 11
@@ -25,7 +38,7 @@
 - 
 ### Added:
 - added 05_sound_to_text tutorial
-- added WavReader to mltu/preprocessors, used to read wav files and convert them to numpy arrays
+- added `WavReader` to `mltu/preprocessors`, used to read wav files and convert them to numpy arrays
 
 
 ## [0.1.7] - 2022-02-03
@@ -35,11 +48,11 @@
 
 ## [0.1.5] - 2022-01-10
 ### Changed
-- seperated CWERMetric to SER and WER Metrics in mltu.metrics, Character/word rate was calculatted in a wrong way
+- seperated `CWERMetric` to `CER` and `WER` Metrics in `mltu.metrics`, Character/word rate was calculatted in a wrong way
 - created @setter for augmentors and transformers in DataProvider, to properlly add augmentors and transformers to the pipeline
 - augmentors and transformers must inherit from `mltu.augmentors.base.Augmentor` and `mltu.transformers.base.Transformer` respectively
 - updated ImageShowCV2 transformer documentation
-- fixed OnnxInferenceModel in mltu.inferenceModels to use CPU even if GPU is available with force_cpu=True flag
+- fixed OnnxInferenceModel in `mltu.inferenceModels` to use CPU even if GPU is available with force_cpu=True flag
 
 ### Added:
 - added RandomSharpen to mltu.augmentors, used for simple image augmentation;